IDIAP MSER implementation
Blatant stealing - but we'll probably release GPL as well so, that's
kind of fair really.

IDIAP MSER is about half the processing time of OpenCV MSER, based on
a tiny little test.

But I don't think we fully resolve the question of why live grid detection
is a little awkward in some places.